feat: SquareContentFactory 추상화 및 파일 분리
